Move BoostAsio.h to SpCore
