+++ cpp/unreal_plugins/SimulationController/Source/SimulationController/RpcServer.h
@@ -11,7 +11,7 @@
 #include <string>
 #include <utility> // std::forward, std::move
 
-#include "SimulationController/BoostAsio.h"
+#include "SpCore/BoostAsio.h"
 #include "SpCore/Rpclib.h"
 
 // This class defines an RPC server in which user-specified functions can be bound to run synchronously in a specific
